+++ imdbinfo/models.py
@@ -33,6 +33,15 @@ class Person(BaseModel):
             job=str(data['knownForJobCategory'])
         )
 
+    @classmethod
+    def from_category(cls, data: dict):
+        return cls(
+            name=data['rowTitle'],
+            id=data['id'],
+            url=f"https://www.imdb.com/name/{data['id']}",
+            job=str(data.get('jobTitle', ''))
+        )
+
 
 class MovieDetail(BaseModel):
     imdbId: str
@@ -47,6 +56,7 @@ class MovieDetail(BaseModel):
     certificates: Dict[str, Tuple[str, str]] = {}
     directors: List[Person] = []
     cast: List[Person] = []
+    stars: List[Person] = []
     year: Optional[int] = None
     duration: Optional[int] = None
     country_codes: List[str] = []
@@ -71,6 +81,7 @@ class MovieDetail(BaseModel):
     summaries: List[str] = []
     synopses: List[str] = []
     production: List[str] = []
+    categories: Dict[str, List[Person]] = {}
 
 class MovieInfo(BaseModel):